Bipartito, rebasado con greedy, funcionando.
